Remove getBoardWithMatchingVidPidFromCloud and make a new class BoardCloudResolver
