+++ arduino-core/src/processing/app/Platform.java
@@ -23,11 +23,6 @@
 package processing.app;
 
 import cc.arduino.packages.BoardPort;
-import cc.arduino.utils.network.HttpConnectionManager;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import processing.app.debug.TargetBoard;
 import processing.app.debug.TargetPackage;
 import processing.app.debug.TargetPlatform;
@@ -36,9 +31,6 @@ import processing.app.legacy.PConstants;
 import javax.swing.*;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.*;
 
 import static processing.app.I18n.tr;
@@ -60,8 +52,7 @@ import static processing.app.I18n.tr;
  * know if name is proper Java package syntax.)
  */
 public class Platform {
-  private static Logger log = LogManager.getLogger(Platform.class);
-
+  // DO NOT USE log4j here otherwise the root path of the logs will no be initialize correctly
   /**
    * Set the default L & F. While I enjoy the bounty of the sixteen possible
    * exception types that this UIManager method might throw, I feel that in
@@ -185,36 +176,6 @@ public class Platform {
     return list;
   }
 
-  public synchronized void getBoardWithMatchingVidPidFromCloud(String vid, String pid) {
-    // this method is less useful in Windows < WIN10 since you need drivers to be already installed
-    ObjectMapper mapper = new ObjectMapper();
-    mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-    try {
-      URL jsonUrl = new URL(String.format("https://builder.arduino.cc/builder/v1/boards/0x%s/0x%s", vid, pid));
-
-      final HttpURLConnection httpConnection = new HttpConnectionManager(jsonUrl)
-        .makeConnection();
-      int code = httpConnection.getResponseCode();
-      if (code == 404) {
-        log.warn("Fail to get the Vid Pid information from the builder response code={}", code);
-        return;
-      }
-      InputStream is = httpConnection.getInputStream();
-      BoardCloudAPIid board = mapper.readValue(is, BoardCloudAPIid.class);
-      log.info("Board info from the cloud {}", board);
-      // Launch a popup with a link to boardmanager#board.getName()
-      // replace spaces with &
-      String realBoardName = board.getName().replaceAll("\\(.*?\\)", "").trim();
-      String boardNameReplaced = realBoardName.replaceAll(" ", "&");
-      String message = I18n.format(tr("{0}Install this package{1} to use your {2} board"), "<a href=\"http://boardsmanager/all#" + boardNameReplaced + "\">", "</a>", realBoardName);
-      BaseNoGui.setBoardManagerLink(message);
-    } catch (Exception e) {
-      // No connection no problem, fail silently
-      //e.printStackTrace();
-
-    }
-  }
-
   public synchronized Map<String, Object> resolveDeviceByVendorIdProductId(String serial, Map<String, TargetPackage> packages) {
     String vid_pid_iSerial = resolveDeviceAttachedToNative(serial);
     for (TargetPackage targetPackage : packages.values()) {
@@ -254,56 +215,6 @@ public class Platform {
       }
     }
     return null;
-  }
-
-  public static class BoardCloudAPIid {
-
-    private String fqbn;
-    private String name;
-    private String architecture;
-    private String id;
-
-    public String getName() {
-      return name;
-    }
-
-    public String getFqbn() {
-      return fqbn;
-    }
-
-    public String getArchitecture() {
-      return architecture;
-    }
-
-    public String getId() {
-      return id;
-    }
-
-    public void setName(String tmp) {
-      name = tmp;
-    }
-
-    public void setFqbn(String fqbn) {
-      this.fqbn = fqbn;
-    }
-
-    public void setArchitecture(String tmp) {
-      architecture = tmp;
-    }
-
-    public void setId(String tmp) {
-      id = tmp;
-    }
-
-    @Override
-    public String toString() {
-      return "BoardCloudAPIid{" +
-        "name='" + name + '\'' +
-        ", fqbn='" + fqbn + '\'' +
-        ", architecture='" + architecture + '\'' +
-        ", id='" + id + '\'' +
-        '}';
-    }
   }
 
   public String resolveDeviceByBoardID(Map<String, TargetPackage> packages, String boardId) {
